LASERWAR pin scheme, PWM prototype
Turned out that LASERWAR's pinout is incompatible with Poligon64 and LSD which can be connected directly to Arduino. Laserwar uses common positive for power that is connected to the +5Vout, while two rest are LED and IR negatives, therefore outputs. They're meant to be connected through the pull down transistors to drive.